Use size_t for open_memstrem

+++ lib/type.c
@@ -538,7 +538,7 @@ static long varlink_type_print(VarlinkType *type,
 const char *varlink_type_get_typestring(VarlinkType *type) {
         FILE *stream = NULL;
         _cleanup_(freep) char *string = NULL;
-        unsigned long size;
+        size_t size;
 
         if (type->typestring)
                 return type->typestring;
